make output/webview delegate internal
