+++ src/vs/workbench/contrib/notebook/browser/view/output/outputRenderer.ts
@@ -6,11 +6,11 @@
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { OutputRendererRegistry } from 'vs/workbench/contrib/notebook/browser/view/output/rendererRegistry';
 import { onUnexpectedError } from 'vs/base/common/errors';
-import { ICellOutputViewModel, INotebookDelegateForOutput, IRenderOutput, RenderOutputType } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
+import { ICellOutputViewModel, IRenderOutput, RenderOutputType } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
 import { URI } from 'vs/base/common/uri';
 import { dispose } from 'vs/base/common/lifecycle';
 import { localize } from 'vs/nls';
-import { IOutputTransformContribution } from 'vs/workbench/contrib/notebook/browser/view/notebookRenderingCommon';
+import { INotebookDelegateForOutput, IOutputTransformContribution } from 'vs/workbench/contrib/notebook/browser/view/notebookRenderingCommon';
 
 export class OutputRenderer {
